Move side arrows to their own file
