+++ src/components/GameTile.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { DungeonTiles } from "@/dungeon_tiles";
 import { getItem } from "@/items";
-import type { Item, Tile } from "game-core/pkg/wasm";
+import type { Item, Tile } from "game-core/pkg";
 import { computed } from "vue";
 
 const props = defineProps<{
